pkg-rabbitmq/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "outamation_pkg_rabbitmq"
version = "0.1.0"
description = "A robust asynchronous RabbitMQ client utility package built on aio-pika for Outamation projects."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "Proprietary" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Topic :: Communications",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# Core runtime dependencies for the RabbitMQ package
dependencies = [
    "aio-pika==9.5.7",
    "python-dotenv==1.1.1",
    "aiohttp==3.12.15",
]

[project.urls]
Homepage = "https://github.com/outamation/core-libs-python"
Repository = "https://github.com/outamation/core-libs-python/tree/main/pkg-rabbitmq"

[tool.setuptools.packages.find]
where = ["src"]  # Look for the source code in the 'src' directory
